refactor: move parameter server integration tests to new folder
